Extract common collection helpers to a module to avoid too much inheritance
[Carlos Antonio da Silva + Rafael Mendonça França]
